% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{assign_cluster}
\alias{assign_cluster}
\alias{assign_cluster.default}
\alias{assign_cluster.kmeans_cluster}
\title{Cluster Assignment of Documents/Text Elements}
\usage{
assign_cluster(x, ...)

\method{assign_cluster}{default}(x, ...)

\method{assign_cluster}{kmeans_cluster}(x, ...)
}
\arguments{
\item{x}{a \code{kmeans_cluster} object.}

\item{\ldots}{ignored.}
}
\value{
Returns an \code{assign_cluster} object; a named vector of cluster
assignments with documents as names.  The object also contains the original
\code{data_storage} object.
}
\description{
Cluster assignment of documents/text elements.
}
\examples{
library(dplyr)

x <- with(
    presidential_debates_2012,
    data_store(dialogue, paste(person, time, sep = "_"))
)

}

