% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_removed.R
\name{get_removed}
\alias{get_removed}
\alias{get_removed.default}
\alias{get_removed.kmeans_cluster}
\title{Get a Text Stored in a \code{kmeans_cluster} Object}
\usage{
get_removed(x, ...)

\method{get_removed}{default}(x, ...)

\method{get_removed}{kmeans_cluster}(x, ...)
}
\arguments{
\item{x}{A \code{\link[hclustext]{kmeans_cluster}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector of text strings.
}
\description{
Extract the text supplied to the
\code{\link[hclustext]{kmeans_cluster}} object.
}
\examples{
library(dplyr)

presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    kmeans_cluster() \%>\%
    get_removed()
}

