% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\alias{get_text.default}
\alias{get_text.kmeans_cluster}
\title{Get a Text Stored in Various Objects}
\usage{
get_text(x, ...)

\method{get_text}{kmeans_cluster}(x, ...)

\method{get_text}{default}(x, ...)
}
\arguments{
\item{x}{A \code{\link[kmeanstext]{kmeans_cluster}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector or list of text strings.
}
\description{
Extract the text supplied to the
\code{\link[kmeanstext]{kmeans_cluster}} object.
}
\examples{
library(dplyr)

}

