% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cluster.R
\name{kmeans_cluster}
\alias{kmeans_cluster}
\alias{kmeans_cluster.data_store}
\title{Fit a Kmeans Cluster}
\usage{
kmeans_cluster(x, k = approx_k(get_dtm(x)), ...)

\method{kmeans_cluster}{data_store}(x, k = approx_k(get_dtm(x)), ...)
}
\arguments{
\item{x}{A data type (e.g., \code{\link[tm]{DocumentTermMatrix}} or
\code{\link[tm]{TermDocumentMatrix}}).}

\item{k}{The number of clusters.  Defaults to use \code{approx_k} of the
\code{\link[tm]{DocumentTermMatrix}} produced by \code{data_storage}.}

\item{\ldots}{Other arguments passed to \code{\link[stats]{kmeans}}.}
}
\value{
Returns an object of class \code{"kmeans"}.
}
\description{
Fit a kmeans cluster to text data.  Prior to distance measures being
calculated the tf-idf (see \code{\link[tm]{weightTfIdf}}) is applied to the
\code{\link[tm]{DocumentTermMatrix}}.
}
\examples{
library(dplyr)
}

