% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeanstext-package.R
\docType{package}
\name{kmeanstext}
\alias{kmeanstext}
\alias{kmeanstext-package}
\alias{package-kmeanstext}
\title{Optimized kmeans Clustering for Text Data}
\description{
Optimized tools for clustering text data via kmeans clustering.
}

